requirements.txt
flask==3.0.2
firebase-admin==6.4.0
requests==2.31.0
python-dotenv==1.0.1
flask-login==0.6.3
gunicorn==21.2.0 